+++ src/index.js
@@ -1,17 +1,23 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+
+import {BrowserRouter, Routes, Route,} from "react-router-dom";
+
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+
+import PokemonDetail from './PokemonDetail';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  <BrowserRouter>
+    <Routes>
+      <Route path="/" element={<App />} />
+      <Route path="detail/:id" element={<PokemonDetail />} />
+      <Route path="*" element={<main><p>404</p></main>}
+    />
+    </Routes>
+  </BrowserRouter>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+